requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pillow
python-multipart